backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.116.1",
    "google-generativeai>=0.8.5",
    "pydantic-settings>=2.10.1",
    "pygad>=3.5.0",
    "sqlmodel>=0.0.24",
]
